ref: move to readme.go analyzer
