+++ pkg/analysis/passes/readme/readme.go
@@ -3,6 +3,7 @@ package readme
 import (
 	"os"
 	"path/filepath"
+	"regexp"
 	"strings"
 
 	"github.com/grafana/plugin-validator/pkg/analysis"
@@ -11,6 +12,7 @@ import (
 
 var (
 	missingReadme = &analysis.Rule{Name: "missing-readme", Severity: analysis.Error}
+	readmeComment = &analysis.Rule{Name: "readme-comment", Severity: analysis.Warning}
 )
 
 var Analyzer = &analysis.Analyzer{
@@ -42,5 +44,22 @@ func run(pass *analysis.Pass) (interface{}, error) {
 		missingReadme.Severity = analysis.OK
 		pass.ReportResult(pass.AnalyzerName, missingReadme, "README.md: exists", "")
 	}
+
+	readmeContent := string(b)
+
+	commentRegex := `<!--(.*?)-->`
+	re := regexp.MustCompile(commentRegex)
+
+	// No need find all
+	comment := re.FindString(readmeContent)
+
+	if len(comment) > 0 {
+		pass.ReportResult(pass.AnalyzerName, readmeComment, "README.md contains comment(s).", "")
+	}
+	if readmeComment.ReportAll {
+		readmeComment.Severity = analysis.OK
+		pass.ReportResult(pass.AnalyzerName, readmeComment, "README.md contains no comment(s).", "")
+	}
+
 	return b, nil
 }
