+++ pkg/analysis/passes/readme/testdata/comment/README.md
@@ -1,0 +1,5 @@
+# Comment
+
+<!-- Some comment -->
+
+Normal text.
